Convert keypoint label to int before converting to str in draw_keypoints

+++ court_line_detector/court_line_detector.py
@@ -50,7 +50,7 @@ class CourtLineDetector:
             y = int(keypoints[i + 1])
             
             cv2.circle(image, (x, y), 5, (0, 0, 255), -1)
-            cv2.putText(image, str(i / 2), (x, y - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
+            cv2.putText(image, str(int(i / 2)), (x, y - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
 
         return image
